+++ backend/src/main/java/com/debuglog/repository/ResolutionRepository.java
@@ -4,6 +4,6 @@ import com.debuglog.model.Resolution;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
-public interface ResolutionRepository extends JpaRepository<Resolution, Long> {
+public interface ResolutionRepository extends JpaRepository<Resolution, Integer> {
     Optional<Resolution> findTopByErrorHashAndConfirmedTrue(String errorHash);
 }
