+++ backend/src/main/java/com/debuglog/service/KnowledgeBaseService.java
@@ -8,31 +8,31 @@ import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
-public class KnowledgeBaseService {
+public class KnowledgeBaseService { 
 
     private final ResolutionRepository repository;
 
     public Optional<Resolution> findByHash(String errorHash) {
         return repository.findTopByErrorHashAndConfirmedTrue(errorHash)
-                .map(r -> {
-                    r.setUsedCount(r.getUsedCount() + 1);
-                    return repository.save(r);
-                });
+            .map(r -> {
+                r.setUsedCount(r.getUsedCount() + 1);
+                return repository.save(r);
+            });
     }
 
     public Resolution save(String errorHash, String errorType,
                            String pattern, String solution, boolean confirmed) {
         return repository.save(Resolution.builder()
-                .errorHash(errorHash)
-                .errorType(errorType)
-                .pattern(pattern)
-                .solution(solution)
-                .confirmed(confirmed)
-                .usedCount(0)
-                .build());
+            .errorHash(errorHash)
+            .errorType(errorType)
+            .pattern(pattern)
+            .solution(solution)
+            .confirmed(confirmed)
+            .usedCount(0)
+            .build());
     }
 
-    public void confirm(Long id) {
+    public void confirm(Integer id) {
         repository.findById(id).ifPresent(r -> {
             r.setConfirmed(true);
             repository.save(r);
